Refactor add base for role update parameters and dto's
