+++ Core.ApplicationServices/Core.ApplicationServices.csproj
@@ -210,6 +210,7 @@
     <Compile Include="Model\Organizations\Write\MasterDataRoles\ContactPersonUpdateParameters.cs" />
     <Compile Include="Model\Organizations\Write\MasterDataRoles\DataProtectionAdvisorUpdateParameters.cs" />
     <Compile Include="Model\Organizations\Write\MasterDataRoles\DataResponsibleUpdateParameters.cs" />
+    <Compile Include="Model\Organizations\Write\MasterDataRoles\OrganizationMasterDataRoleUpdateParameters.cs" />
     <Compile Include="Model\Organizations\Write\MasterDataRoles\OrganizationMasterDataRolesUpdateParameters.cs" />
     <Compile Include="Model\Organizations\Write\OrganizationUnitUpdateParameters.cs" />
     <Compile Include="Model\Organizations\Write\OrganizationMasterDataUpdateParameters.cs" />
