+++ Core.ApplicationServices/Model/Organizations/Write/MasterDataRoles/ContactPersonUpdateParameters.cs
@@ -2,10 +2,8 @@
 
 namespace Core.ApplicationServices.Model.Organizations.Write.MasterDataRoles;
 
-public class ContactPersonUpdateParameters
+public class ContactPersonUpdateParameters: OrganizationMasterDataRoleUpdateParameters
 {
-    public OptionalValueChange<string> Name { get; set; }
     public OptionalValueChange<string> LastName { get; set; }
     public OptionalValueChange<string> PhoneNumber { get; set; }
-    public OptionalValueChange<string> Email { get; set; }
 }
