+++ Core.ApplicationServices/Model/Organizations/Write/MasterDataRoles/DataProtectionAdvisorUpdateParameters.cs
@@ -2,11 +2,9 @@
 
 namespace Core.ApplicationServices.Model.Organizations.Write.MasterDataRoles;
 
-public class DataProtectionAdvisorUpdateParameters
+public class DataProtectionAdvisorUpdateParameters: OrganizationMasterDataRoleUpdateParameters
 {
-    public OptionalValueChange<string> Name { get; set; }
     public OptionalValueChange<string> Cvr { get; set; }
     public OptionalValueChange<string> Phone { get; set; }
     public OptionalValueChange<string> Address { get; set; }
-    public OptionalValueChange<string> Email { get; set; }
 }
