+++ Core.ApplicationServices/Model/Organizations/Write/MasterDataRoles/OrganizationMasterDataRoleUpdateParameters.cs
@@ -1,0 +1,10 @@
+﻿using Core.ApplicationServices.Model.Shared;
+
+namespace Core.ApplicationServices.Model.Organizations.Write.MasterDataRoles
+{
+    public class OrganizationMasterDataRoleUpdateParameters
+    {
+        public OptionalValueChange<string> Name { get; set; }
+        public OptionalValueChange<string> Email { get; set; }
+    }
+}
