+++ Presentation.Web/Models/API/V2/Internal/Request/Organizations/DataProtectionAdvisorRequestDTO.cs
@@ -5,12 +5,10 @@ using System.Web;
 
 namespace Presentation.Web.Models.API.V2.Internal.Request.Organizations
 {
-    public class DataProtectionAdvisorRequestDTO
+    public class DataProtectionAdvisorRequestDTO: OrganizationMasterDataRoleRequestDTO
     {
-        public string Name { get; set; }
         public string Cvr { get; set; }
         public string Phone { get; set; }
         public string Address { get; set; }
-        public string Email { get; set; }
     }
 }
