+++ Presentation.Web/Models/API/V2/Internal/Request/Organizations/DataResponsibleRequestDTO.cs
@@ -5,12 +5,10 @@ using System.Web;
 
 namespace Presentation.Web.Models.API.V2.Internal.Request.Organizations
 {
-    public class DataResponsibleRequestDTO
+    public class DataResponsibleRequestDTO: OrganizationMasterDataRoleRequestDTO
     {
-        public string Name { get; set; }
         public string Cvr { get; set; }
         public string Phone { get; set; }
         public string Address { get; set; }
-        public string Email { get; set; }
     }
 }
