+++ Presentation.Web/Models/API/V2/Internal/Request/Organizations/OrganizationMasterDataRoleRequestDTO.cs
@@ -5,9 +5,9 @@ using System.Web;
 
 namespace Presentation.Web.Models.API.V2.Internal.Request.Organizations
 {
-    public class ContactPersonRequestDTO: OrganizationMasterDataRoleRequestDTO
+    public class OrganizationMasterDataRoleRequestDTO
     {
-        public string LastName { get; set; }
-        public string PhoneNumber { get; set; }
+        public string Name { get; set; }
+        public string Email { get; set; }
     }
 }
