+++ Presentation.Web/Models/API/V2/Internal/Response/Organizations/ContactPersonResponseDTO.cs
@@ -6,12 +6,9 @@ using Presentation.Web.Models.API.V2.Response.Organization;
 
 namespace Presentation.Web.Models.API.V2.Internal.Response.Organizations
 {
-    public class ContactPersonResponseDTO
+    public class ContactPersonResponseDTO: OrganizationMasterDataRoleResponseDTO
     {
-            public int Id { get; set; }
-            public string Name { get; set; }
             public string LastName { get; set; }
             public string PhoneNumber { get; set; }
-            public string Email { get; set; }
         }
     }
