+++ Presentation.Web/Models/API/V2/Internal/Response/Organizations/DataProtectionAdvisorResponseDTO.cs
@@ -5,13 +5,10 @@ using System.Web;
 
 namespace Presentation.Web.Models.API.V2.Internal.Response.Organizations
 {
-    public class DataProtectionAdvisorResponseDTO
+    public class DataProtectionAdvisorResponseDTO: OrganizationMasterDataRoleResponseDTO
     {
-        public string Name { get; set; }
         public string Cvr { get; set; }
         public string Phone { get; set; }
         public string Address { get; set; }
-        public string Email { get; set; }
-        public int Id { get; set; }
     }
 }
