+++ Presentation.Web/Models/API/V2/Internal/Response/Organizations/OrganizationMasterDataRoleResponseDTO.cs
@@ -1,0 +1,14 @@
+﻿using System;
+using System.Collections.Generic;
+using System.Linq;
+using System.Web;
+
+namespace Presentation.Web.Models.API.V2.Internal.Response.Organizations
+{
+    public class OrganizationMasterDataRoleResponseDTO
+    {
+        public string Name { get; set; }
+        public string Email { get; set; }
+        public int Id { get; set; }
+    }
+}
